feat: move /components/modals/ to modals/
